+++ src/edu/yu/einstein/genplay/core/writer/binListWriter/ConcatenateBinListWriter.java
@@ -14,7 +14,7 @@
  *
  *     You should have received a copy of the GNU General Public License
  *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *     
+ * 
  *     Authors:	Julien Lajugie <julien.lajugie@einstein.yu.edu>
  *     			Nicolas Fourel <nicolas.fourel@einstein.yu.edu>
  *     Website: <http://genplay.einstein.yu.edu>
@@ -27,8 +27,11 @@ import java.io.FileWriter;
 import java.io.IOException;
 
 import edu.yu.einstein.genplay.core.chromosome.Chromosome;
+import edu.yu.einstein.genplay.core.enums.AlleleType;
 import edu.yu.einstein.genplay.core.list.binList.BinList;
 import edu.yu.einstein.genplay.core.manager.project.ProjectManager;
+import edu.yu.einstein.genplay.core.multiGenome.utils.FormattedMultiGenomeName;
+import edu.yu.einstein.genplay.core.multiGenome.utils.ShiftCompute;
 import edu.yu.einstein.genplay.core.writer.Writer;
 import edu.yu.einstein.genplay.exception.BinListDifferentWindowSizeException;
 import edu.yu.einstein.genplay.gui.statusBar.Stoppable;
@@ -45,8 +48,11 @@ public class ConcatenateBinListWriter implements Writer, Stoppable {
 	private final BinList[] 		binListArray;			// array of the BinList to concatenate
 	private final String[] 			nameArray;				// name of the BinLists
 	private final File				outputFile;				// file where to write the concatenation
-	private boolean					needsToBeStopped = false;// true if the writing need to be stopped 
-	
+	private boolean					needsToBeStopped = false;// true if the writing need to be stopped
+
+	private String					fullGenomeName;		// the genome name (multi genome project only)
+	private AlleleType				allele;				// the allele type (multi genome project only)
+
 
 	/**
 	 * Creates an instance of {@link ConcatenateBinListWriter}
@@ -57,7 +63,14 @@ public class ConcatenateBinListWriter implements Writer, Stoppable {
 	public ConcatenateBinListWriter(BinList[] binListArray, String[] nameArray, File outputFile) {
 		this.binListArray = binListArray;
 		this.nameArray = nameArray;
-		this.outputFile = outputFile;		
+		this.outputFile = outputFile;
+	}
+
+
+	@Override
+	public void setMultiGenomeCoordinateSystem (String genome, AlleleType allele) {
+		this.fullGenomeName = genome;
+		this.allele = allele;
 	}
 
 
@@ -65,12 +78,13 @@ public class ConcatenateBinListWriter implements Writer, Stoppable {
 	 * Concatenates and saves a list of {@link BinList} in a file
 	 * @throws IOException
 	 * @throws BinListDifferentWindowSizeException
-	 * @throws InterruptedException 
+	 * @throws InterruptedException
 	 */
+	@Override
 	public void write() throws IOException, BinListDifferentWindowSizeException, InterruptedException {
 		if (binListArray.length > 0) {
-			
-			// check if the BinList all have the same bin size 
+
+			// check if the BinList all have the same bin size
 			int binSize = binListArray[0].getBinSize();
 			for (BinList currentList: binListArray) {
 				if (currentList.getBinSize() != binSize) {
@@ -80,6 +94,8 @@ public class ConcatenateBinListWriter implements Writer, Stoppable {
 
 			BufferedWriter writer = null;
 			try {
+				boolean isMultiGenome = ProjectManager.getInstance().isMultiGenomeProject() && (fullGenomeName != null) && (allele != null);
+
 				// try to create a output file
 				writer = new BufferedWriter(new FileWriter(outputFile));
 
@@ -99,28 +115,42 @@ public class ConcatenateBinListWriter implements Writer, Stoppable {
 					writer.write("\t" + nameArray[i]);
 				}
 				writer.newLine();
-				
-				for (Chromosome currentChromo: ProjectManager.getInstance().getProjectChromosome()) {					
-					int binCount = currentChromo.getLength() / binSize + 1; 
+
+				for (Chromosome currentChromosome: ProjectManager.getInstance().getProjectChromosome()) {
+					int currentChromosomeSize = currentChromosome.getLength();
+					int binCount = (currentChromosomeSize / binSize) + 1;
 					int j = 0;
 					while (j < binCount) {
-						writer.write(currentChromo + "\t" + (j * binSize) + "\t" + ((j + 1) * binSize));
-						for (BinList currentBinList: binListArray) {
-							// if the operation need to be stopped we close the writer and delete the file 
-							if (needsToBeStopped) {
-								writer.close();
-								outputFile.delete();
-								throw new InterruptedException();
-							}							
-							if ((currentBinList.get(currentChromo) != null) && (j < currentBinList.size(currentChromo))) {
-								writer.write("\t" + currentBinList.get(currentChromo, j));
-							} else {
-								writer.write("\t0.0");
-							}
+						int start = j * binSize;
+						int stop = start + binSize;
+						if (stop > currentChromosomeSize) {
+							stop = currentChromosomeSize;
 						}
-						writer.newLine();
-						j++;						
-					}					
+
+						if (isMultiGenome) {
+							start = ShiftCompute.getPosition(FormattedMultiGenomeName.META_GENOME_NAME, allele, start, currentChromosome, fullGenomeName);
+							stop = ShiftCompute.getPosition(FormattedMultiGenomeName.META_GENOME_NAME, allele, stop, currentChromosome, fullGenomeName);
+						}
+
+						if ((start != -1) && (stop != -1)) {
+							writer.write(currentChromosome + "\t" + start + "\t" + stop);
+							for (BinList currentBinList: binListArray) {
+								// if the operation need to be stopped we close the writer and delete the file
+								if (needsToBeStopped) {
+									writer.close();
+									outputFile.delete();
+									throw new InterruptedException();
+								}
+								if ((currentBinList.get(currentChromosome) != null) && (j < currentBinList.size(currentChromosome))) {
+									writer.write("\t" + currentBinList.get(currentChromosome, j));
+								} else {
+									writer.write("\t0.0");
+								}
+							}
+							writer.newLine();
+						}
+						j++;
+					}
 				}
 			} finally {
 				if (writer != null) {
@@ -136,6 +166,6 @@ public class ConcatenateBinListWriter implements Writer, Stoppable {
 	 */
 	@Override
 	public void stop() {
-		needsToBeStopped = true;		
+		needsToBeStopped = true;
 	}
 }
